merge sonarsource/sonarqube to michaeljesse73/sonarqube

+++ server/sonar-db-migration/src/main/java/org/sonar/server/platform/db/migration/step/BaseSqlStatement.java
@@ -95,7 +95,7 @@ class BaseSqlStatement<CHILD extends SqlStatement> implements SqlStatement<CHILD
   @SuppressWarnings("unchecked")
   public CHILD setDouble(int columnIndex, @Nullable Double value) throws SQLException {
     if (value == null) {
-      pstmt.setNull(columnIndex, Types.DECIMAL);
+      pstmt.setNull(columnIndex, Types.DOUBLE);
     } else {
       pstmt.setDouble(columnIndex, value);
     }
